Reuse common constants and types
