+++ src/components/slider/slider-types.ts
@@ -5,3 +5,14 @@ export interface PointerPosition {
     clientX: number;
     clientY: number;
 }
+export type KeyboardHandler = (event: React.KeyboardEvent<HTMLElement>) => void;
+export type MouseHandler = (
+    event: React.MouseEvent<HTMLElement> | MouseEvent,
+    sliderArea: HTMLElement,
+    focusableElement: HTMLElement
+) => void;
+export type TouchHandler = (
+    event: React.TouchEvent<HTMLElement> | TouchEvent,
+    sliderArea: HTMLElement,
+    focusableElement: HTMLElement
+) => void;
